requirements.txt
pyyaml
iso8601==0.1.4
python-dateutil==1.5
ipython
lxml>=2.2
pytz
cssselect
scrapelib>=0.10.0,<1.0.0
mechanize
BeautifulSoup4
mock
inflection
fs
xmltodict
